+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Added
   - Support for downloadong derivatives as glb (glTF binary format) with Draco compression
   - Default sorting of OSS and webhook entries in the tree view.
+- Fixed
+  - When downloading an SVF and certain assets cannot be accessed, issue a warning and continue
 
 ## [1.3.0] - 2019-10-17